requirements.txt
streamlit
pandas
python-dotenv
openai
sounddevice
scipy
numpy
ibm-watson
watchdog
flask
pygame
python-dateutil
pytz